test: 100%

+++ storage/redis.js
@@ -66,39 +66,68 @@ class StorageRedis extends StorageInterface {
 
   async remove (key) {
     this.log.debug({ msg: 'acd/storage/redis.remove', key })
-
-    this.store.del(key)
+    try {
+      this.store.del(key)
     // TODO remove key in references? do it lazy/gc?
+    } catch (err) {
+      this.log.error({ msg: 'acd/storage/redis.remove error', err, key })
+    }
   }
 
   async invalidate (references) {
     this.log.debug({ msg: 'acd/storage/redis.invalidate', references })
+
+    try {
     // TODO can nested loops be avoided?
-    for (let i = 0; i < references.length; i++) {
-      const reference = references[i]
-      // TODO pipeline?
-      const keys = await this.store.smembers(reference)
-      this.log.debug({ msg: 'acd/storage/redis.invalidate got keys to be invalidated', keys })
-      if (!keys || keys.length < 1) {
-        continue
-      }
-      for (let j = 0; j < keys.length; j++) {
+      for (let i = 0; i < references.length; i++) {
+        const reference = references[i]
+        // TODO pipeline?
+        const keys = await this.store.smembers(reference)
+        this.log.debug({ msg: 'acd/storage/redis.invalidate got keys to be invalidated', keys })
+        if (!keys || keys.length < 1) {
+          continue
+        }
+        for (let j = 0; j < keys.length; j++) {
         // TODO can be done in 1 query? pipeline?
-        this.log.debug({ msg: 'acd/storage/redis.del key' + keys[j] })
-        // TODO! if not store key => this._store.sdel(reference, key)
-        await this.store.del(keys[j])
+          this.log.debug({ msg: 'acd/storage/redis.del key' + keys[j] })
+          await this.store.del(keys[j])
+        }
       }
+    // TODO update references removing deleted keys? gc?
+    } catch (err) {
+      this.log.error({ msg: 'acd/storage/redis.invalidate error', err, references })
     }
-    // TODO update references removing deleted keys?
   }
 
   async clear (name) {
-    await this.store.flushall()
-    // TODO remove keys starts with name
+    this.log.debug({ msg: 'acd/storage/redis.clear', name })
+
+    try {
+      if (!name) {
+        await this.store.flushall()
+        return
+      }
+
+      const keys = await this.store.keys(name + '*')
+      this.log.debug({ msg: 'acd/storage/redis.clear keys', keys })
+
+      // TODO pipeline
+      const tasks = []
+      for (let i = 0; i < keys.length; i++) {
+        tasks.push(this.store.del(keys[i]))
+      }
+      await Promise.all(tasks)
+    } catch (err) {
+      this.log.error({ msg: 'acd/storage/redis.clear error', err, name })
+    }
   }
 
   async refresh () {
-    await this.store.flushall()
+    try {
+      await this.store.flushall()
+    } catch (err) {
+      this.log.error({ msg: 'acd/storage/redis.refresh error', err })
+    }
   }
 }
